+++ src/owl/core/owl_ndarray_transpose_impl.c
@@ -1,0 +1,4 @@
+/*
+ * OWL - an OCaml numerical library for scientific computing
+ * Copyright (c) 2016-2018 Liang Wang <liang.wang@cl.cam.ac.uk>
+ */